Clean up data model and logic that determines if a resource is in use
The cleanup tasks are fairly critical operations. We're going to be deleting
lots of resources from a production account. While it's not the end of the
world, it's very inconvenient to recover when images or snapshots are unintentionally
deleted. Especially when this is run automagically in a scheduled CI, and everything's
running in an autoscaling group.

Thus, it's worth taking some time to ensure that all of the logic is correct, and the
easiest way to do that is to factor down shared logic to a single definition.