Add `use` and `current` methods to collections.
Calling `use(NAME)` on a collection pins one element as the `current` attribute. This allows consumers, like templates
to reference one element of a collection wthout additional input, e.g.

```ruby
...
config.vm.provider '<%= profile.current.vagrant.local.provider %>' do |local, override|
...
```

in `Vagrantfile.erb` will always reference the selected profile for the build.

To accomplish this, `Attributes` instances now cache the instances of their sub-nodes in a hash, `nodes`. This should
have the added benifit of some improved performance as we're not instantiating new instances of the same wrapper
class for every traversal of a namespace or collection.