Prevents unwanted from being bundled
This change limits the files and directories we include in the gem.

It fixes the situation where you locally build the gem and try to
include in your local bundle(`gem 'builderator', path:
'../builderator'`)

This is the specific error we are trying to prevent:
Gem::InvalidSpecificationException: The gemspec at
/Users/athompson/code/builderator/builderator.gemspec is not valid.
Please fix this gemspec.
The validation error was 'builderator-0.0.1 contains itself
(builderator-0.0.1.gem), check your files list'