Task to manage chefignore file
The aim of this change is to try to detect the presence of the ignore
files/directores in the `chefignore` file and try to add them if they
are missing.

In addition to adding the ignore files to the `chefignore` file we are
displaying a warning to the user suggesting they update their
`chefignore` file in the source repo so we don't have to do this.

Finally, we copy the `chefignore` file to the `.builderator` so that the
`berks` binary can find and use it.

The content we are adding looks like this:

```
.builderator
```

We are no longer nesting vendored cookbooks, so this fixes #88.

```
$ find .builderator -name rapid7-logstash-server

$ find .builderator/cookbooks | wc -l
     813
```