Create security group with tight ssh permissions
By default, if you don't specify a security group, packer will create a
security group with port 22 (ssh) open to the world.  This is
suboptimal, as this level of openness is unnecessary.

Now, if no security group is specified by the Buildfile, builderator
will create one with port 22 only open to the external IP of the system
on which builderator is being run, as determined by Amazon's external ip
lookup service.  After bake is complete (successfully or not, barring
certain exceptions within ruby) the security group is deleted.

This design is similar to packer's default behavior, except that it
locks the security group down to only give the access necessary.