Use microsecond timestamp when creating security group
PhaaS specifically has a job which builds like 10 jobs at once, and it
just so happened that two different jobs tried to create a security
group with the exact same timestamp.  This was surprising enough, but I
would be much more surprised if this race condition still occurs after
this commit.