Retry image search after build complete
Sometimes, because of Amazon's eventual consistency in their API
databases, searching for an image that was just built will return a nil,
rather than an array of images.  This sucks, because it's not an error
from Amazon, and we don't catch it.  So this happens:

.../gems/builderator-2.0.2/lib/builderator/tasks/packer.rb:44:in `block in copy':
undefined method `image_id' for nil:NilClass (NoMethodError)
  from .../gems/builderator-2.0.2/lib/builderator/tasks/packer.rb:41:in `each'
  from .../gems/builderator-2.0.2/lib/builderator/tasks/packer.rb:41:in `copy'
  from .../gems/thor-0.19.1/lib/thor/command.rb:27:in `run'
  from .../gems/thor-0.19.1/lib/thor/invocation.rb:126:in `invoke_command'
  from .../gems/thor-0.19.1/lib/thor.rb:359:in `dispatch'
  from .../gems/thor-0.19.1/lib/thor/invocation.rb:115:in `invoke'
  from .../gems/builderator-2.0.2/lib/builderator/tasks.rb:90:in `image'
  from .../gems/thor-0.19.1/lib/thor/command.rb:27:in `run'
  from .../gems/thor-0.19.1/lib/thor/invocation.rb:126:in `invoke_command'
  from .../gems/thor-0.19.1/lib/thor.rb:359:in `dispatch'
  from .../gems/thor-0.19.1/lib/thor/base.rb:440:in `start'
  from .../gems/builderator-2.0.2/bin/build:4:in `<top (required)>'
  from .../bin/build:23:in `load'
  from .../bin/build:23:in `<main>'
  from .../bin/ruby_executable_hooks:15:in `eval'
  from .../bin/ruby_executable_hooks:15:in `<main>'

This change grabs the NoMethodError by invoking a method (length) that
exists for arrays but not for nil, and retrying the search a few times
with exponential backoff before giving up and failing.